reorganize Operation into its own files
