+++ dbg.hpp
@@ -51,11 +51,3 @@ class Tracee {
     // Inserts a breakpoint at address `addr` in the child process.
     void insert_breakpoint(size_t addr);
 };
-
-class Operation {
-    std::vector<std::string> get_tokenize_command();
-    int execute_command(std::vector<std::string> arguments, Tracee& tracee);
-
-    public:
-    int parse_and_run(Tracee tracee);
-};
